ahhhhhh

+++ HunterLloydsLastDance/colormash.py
@@ -56,7 +56,7 @@ try:
         hsv_image = cv2.cvtColor(color_image, cv2.COLOR_BGR2HSV)
 
         (h, s, v) = cv2.split(hsv_image)
-        s = int(s/2)+127
+        s = s/2+127
         s = np.clip(s, 0, 255)
         hsv_image = cv2.merge([h, s, v])
